Updated static-resource
