+++ packages/server-core/src/recording/recording/recording.ts
@@ -43,7 +43,7 @@ export default (app: Application): void => {
     multi: true
   }
 
-  app.use(recordingPath, new RecordingService(options, app), {
+  app.use(recordingPath, new RecordingService(options), {
     // A list of all methods this service exposes externally
     methods: recordingMethods,
     // You can add additional custom events to be sent to clients here
